Back out some changes to docs Makefiles
Commit ab0e9dbfa76e056f875e969c0d7b6e133ec75431 introduced some changes
to the documentation Makefiles designed to clean-up the process of
deciding which headers get scanned for the docs.

Unfortunately, the gtk-doc Makefile doesn't use HFILE_GLOB for actually
generating the docs -- only for knowing when it needs to redo the
generation.  Because of this, we need to use IGNORE_HFILES or otherwise
we get hundreds of symbols in the *-unused.txt files.

Revert the changes that that commit made to the docs Makefiles (but
leave the generation of the *-public-headers.txt files in place).