GMappedFile: fail when mapping a device file
mmap() fails on zero-sized files, so we previously had a special case to
avoid calling it in that case.  Unfortunately, this had the side effect
of causing us to fail to notice that we were attempting to mmap() a
device node.

Modify the special-casing to only apply in the case that we're dealing
with a normal file.

https://bugzilla.gnome.org/show_bug.cgi?id=659212